finished raycast
missing demo

+++ inc/Toast/Physics/ColliderData.hpp
@@ -9,7 +9,7 @@ namespace physics {
 
 struct ColliderData {
 	double friction = 0.4;
-	glm::vec2 worldPosition;
+	glm::dvec2 worldPosition;
 
 	bool debugNormals = false;
 };
